stronger typing

+++ Isonia/Pipeline/Descriptors/DescriptorWriter.h
@@ -73,7 +73,7 @@ namespace Isonia::Pipeline::Descriptors
 			{
 				write.dstSet = set;
 			}
-			vkUpdateDescriptorSets(pool.device.GetDevice(), writes.size(), writes.data(), 0, nullptr);
+			vkUpdateDescriptorSets(pool.device.GetDevice(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
 		}
 
 	private:
